Fix HOC type definitions
- Fix type signatures of HOCs
- Simplify types for components
