+++ packages/gasket-typescript-tests/package.json
@@ -70,7 +70,7 @@
     "@types/express": "^4.17.13",
     "@types/jest": "^29.0.0",
     "eslint": "^8.7.0",
-    "eslint-config-godaddy-typescript": "^4.0.0",
+    "eslint-config-godaddy-react-typescript": "^4.0.2",
     "eslint-plugin-unicorn": "^44.0.0",
     "jest": "^29.3.1",
     "redux": "^4.0.5",
@@ -79,7 +79,7 @@
   },
   "eslintConfig": {
     "extends": [
-      "godaddy-typescript",
+      "godaddy-react-typescript",
       "plugin:jest/recommended"
     ],
     "plugins": [
